+++ cmd/findo/main.go
@@ -6,7 +6,7 @@ import (
 	"github.com/geekjourneyx/findo/internal/cli"
 )
 
-var version = "1.2.2"
+var version = "1.2.3"
 
 func main() {
 	os.Exit(cli.Run(os.Args[1:], version, os.Stdout, os.Stderr))